Move isin tests into the correct files and add test of error case for categorical DataFrame.
